requirements.txt
absl-py==0.12.0
astunparse==1.6.3
async-generator==1.10
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.5
clang==5.0
cycler==0.10.0
dill==0.3.4
entrypoints==0.3
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gatspy==0.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
importlib-resources==5.3.0
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
neurokit2==0.1.5
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.3
Pillow==8.3.2
pipreqs==0.4.11
promise==2.3
protobuf==3.17.3
py-ecg-detectors==1.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pytz==2021.1
PyWavelets==1.2.0
pywin32==228
pywinpty==0.5.7
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
sklearn==0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.14.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
tensorflow-metadata==1.4.0
termcolor==1.1.0
terminado==0.9.4
threadpoolctl==2.2.0
tqdm==4.62.2
typeguard==2.13.0
urllib3==1.26.6
webencodings==0.5.1
Werkzeug==2.0.1
wfdb==3.4.1
wincertstore==0.2
wrapt==1.12.1
yarg==0.1.9